Added Butterbar component (#354)
* Added Butterbar component

Visually equivalent to the existing components/Butterbar/index.js in the monolith. This one also has the option to switch to a visualization with a border-bottom, as that's the desired design for scenarios where the content just below the butterbar has a white background and/or makes the butterbar inside a visual container.

I verified this looks the same when swapped out in the monolith's SkillPathsButterbar component with this linked. Once this PR goes in I'll mark the original component as deprecated and apply that switch.

* Actually corrected for storageKey