[PREM-38] Update nav bar logos in gamut (#638)
The upper left nav logos on our site do not use gamut! This PR will change that!

- Replace the gamut Pro logo with an updated one (Pro in purple, not mint)
- Add a mono-color Pro logo. The immediate goal is to have an all-white Pro logo to use in dark mode LE.
- Add a premium logo.

Mike K and Yassine have looked at these pretty thoroughly. This is not a YOLO change. We've also done some VQA on my local with these logos rendered in the nav bar in the site, and the initial response is good.