Added basic Modal primitive to Gamut
This is a very simple version of a modal that acts as a controlled component, receiving isOpen as a prop. It's intentionally much simpler than the overused Modal components in the monolith.

Does not yet include animations. I just want to get this out quickly to unblock certificates & any other work that would use it.