feat: added basic Modal primitive to Gamut (#641)
* Added basic Modal primitive to Gamut

This is a very simple version of a modal that acts as a controlled component, receiving isOpen as a prop. It's intentionally much simpler than the overused Modal components in the monolith.

Does not yet include animations. I just want to get this out quickly to unblock certificates & any other work that would use it.

* Lint fixes

* Combined container styles

* Update packages/gamut/src/Modal/__tests__/Modal-test.tsx

* Removed contents

* Fixed compile complaints

* Whitespace added...

* Consolidated React versions