feat(Box): Adds a set of utility Box components using system props
## Overview

Extends our use of the library soon to be known as `variance` to a set of flexible `<Box />` components.  These use a large set of what we call `system props` which you can check out in the story prop table.  The Box components are necessarily a set of pretty much all of them since we want it to be able to do pretty much anything when you need it.

Why in general?
-  Having a generalized utility component that can achieve any task can cut down on custom code and divergent implementations.
- Usage and behavior are tied together.  You can see what it is going to do where it is used, rather than looking up references.  
- Consistent behavior.  Props are locked down to specific scales from our design system.  So you can easily configure your colors and padding sizes with autocomplete. 


Check out the Storybook Page: [here](https://5fcaa48c7d0205176b5a934c--gamut-preview.netlify.app/storybook/?path=/docs/layouts-box--box)

## Additions
- `Box` - Swiss army knife container component that can be configured in pretty much any way imaginable.
- `GridBox` - Box but with `grid` specific props.
- `FlexBox` - Same as grid box but with flex.

Why not all the same components?  
- Grid and Flex overlap on specific prop keys and have different valid values `flex-start` vs `start`. 
- Declarative component names make it easier to infer the use from markup at a glance.

## Changes
- Refactors the original `Box` to become `CardBox` that utilities the Box component to achieve some of its effects.
- Refactors `Badge` to be a configured `Box` instead of its own component.




### PR Checklist

- [x] Related to designs:
- [ ] Related to JIRA ticket: ABC-123
- [x] I have run this code to verify it works
- [ ] This PR includes unit tests for the code change

### Description

<!--
## Merging your changes

When you are ready to merge to main and publish your changes, use the "squash and merge" button in GitHub, and follow the [PR Title Guide](https://github.com/Codecademy/client-modules#pr-title-guide) to format your PR title and write your PR merge description.

**IMPORTANT:** If your PR contains breaking changes, please remember to follow the instructions for breaking changes!
-->