fix(Variance): Add better prop sorting tests
Found some borked behavior with the previous tests just being wrong

Added a better safeguard here for shorthand properties and a more exhaustive test to ensure that the most general shorthands are ordered first to ensure the cascade behaves as expected border is above borderColor to prevent incidental overrides.

This ensures that within the same scope worrying about the cascade is not important.  This is necessary to determine due to composition being ad hoc, having predictability when users may not have full access to cascading beyond merge order.