feat: Add PageSection molecule
As part of [EGG-968](https://codecademy.atlassian.net/browse/EGG-968), I'm creating this abstract `PageSection` component to use throughout the Profile page and later to be used in the dashboard as well. 

I've added some stories to Storybook, so you should be able to see it by searching for `PageSection`. 

I've built it to handle the use-case in the main dashboard as well as everything we'll need for the Profile page:
![image](https://user-images.githubusercontent.com/4298857/114916071-3950e900-9dd9-11eb-9989-34bf525b7884.png)
![image](https://user-images.githubusercontent.com/4298857/114916404-99478f80-9dd9-11eb-8de0-6a03830f32cc.png)