fix: Fix Storybook crashes caused by circular object references on uncalled Icon components
### Context
#### Storybook Behaviors
1. Storybook does some [prettifying of JSX](https://github.com/storybookjs/storybook/blob/next/addons/docs/src/frameworks/react/jsxDecorator.tsx#L117 ) which uses a [recursive object property sort method](https://github.com/algolia/react-element-to-jsx-string/blob/master/src/formatter/sortObject.js#L3). 
2. Args are stringified for presentation in various places.

#### Emotion Behavior 
Uncalled Emotion component functions store name spaced object references that are circular on each emotion instance (`__emotion_real`) which is infinitely deep.  It is unclear what the purpose of this is.

#### Cause of Bug
After introducing system props to Icons - stories that consumed Icon functions as a prop broke due to infinite loops attempting to serialize the function.  

#### Fix
This fix for this is to avoid having icons register as an emotion component and stay a `React.forwardRef` while still having access to emotion styles.  The ideal approach would be to fully replace the `svg` tag with a custom `Svg`  emotion component internal to the forwardRef.  However this is not possible as there are limited ways to customize the actual JSX of `svgr` without a custom babel plugin.

This solution uses the render prop `ClassNames` function exposed by emotion/react to generate a the styles prior and apply them as a classname directly to the intrinsic svg component.  As we no longer have access to the theme context default to all styled components we also need to add a `useTheme()` hook to each Icon to ensure that the system props will not break.

```tsx
<ClassNames>{({ css , cx }) => <svg className={cx(className, css(systemProps(props))} />}</ClassNames>
```

This should have no downstream affects to behavior.

#### Other changes
- Upgraded storybook dependencies just in case to the latest minor version (likely the last before 6.3).