fix(variance): Fix static css mutability for `css` [GM-235]
Some how I figured out a way to throw all those object mutability lessons out the window and Object.assign theme to a a nested of object literal reference :).  This meant the theme would be added to the previous static object and returned as CSS even if the parser ignored it. 

Normally this wouldn't be an issue if you called `system.css` with an object literal as the response would be cached from the first call - preventing the mutability issue: 
```tsx
system.css({
  m: 4,
  '&:hover': {
     bg: 'blue',
  }
})
```

However if you used the object between multiple `css` functions the select configuration would be mutated by the first call the create the cache and causing the second css function to attempt to include theme as styles:
```tsx
const hoverStyles = {
  '&:hover': {
     bg: 'blue',
  }
}. as const;

// Works as expected and caches correctly
const buttonStyles = system.css({
  // stuff
  ...hoverStyles
});

// Whole bunch of pain
const linkStyles =  = system.css({
  // other stuff 
  ...hoverStyles
});
```

This fixes the mutability issue by adding a spread where the unsafe assign was used, like normal.

And by adding `theme` to the filter list if it is somehow passed manually.