fix(Flyout): change how we're setting full height
Changes `height: 100vh` to `bottom: 0` within `Flyout` to fix bug on mobile browsers where the bottom of the `Flyout` was getting covered by the native browser navigation on the bottom.

You can test this by comparing the fix (https://portal-app-pr-176.dev-eks.codecademy.com/catalog) to the current behavior on staging (https://portal-app.staging-eks.codecademy.com/catalog) on your actual mobile device. Click the "Catalog Menu" button at the top of the page to trigger the Flyout and confirm that you are now able to fully see and click on the full catalog and sorting quiz links at the bottom.