feat(FloatingCard): Adds shadow offset direction and pattern
This FloatingCard feature adds a shadow prop to optionally change the direction of the shadow when needed. By default, the shadow is set to the bottom-left (as it were before) and now the pattern of the shadow can also be changed to a pattern that exists in the gamut-pattern kit.