refactor(LayoutMenu): text wrapping support for section items and link
The titles of nav items within the `LayoutMenu` currently do not wrap to the next line inside their containers. If title lengths are really long and beyond the width of its container, the text would overflow because we've set the width of `Anchor` within `SectionItemLink` to be "max-width." To ensure that we maintain focus states on tab _AND_ the section item/links wrap to the next line for the occasional long title, this change will set the widths to "fit-content" to allow text wrapping.