fix(GlobalFooter): don't display anchors if no click handler
In a previous hackathon I added a fun easter egg to the footer on the logged in dashboard (clicking on different parts of "Made in NYC..." open p5 animation overlays. But the anchor styling is applied to all pages even though only the logged in dashboard actually passes in the click handler. This PR will render static text when there's no click handler.