fix(HorizontalScrollMenu): polyfill IntersectionObserver
Around 1% of people are using browsers that don't support IntersectionObserver.  It's causing a lot of errors that have surfaced in datadog monitoring.

I tried dynamically loading the polyfill only if window.IntersectionObserver is undefined, but I was running into issues.  I also tried using `react-intersection-observer` with Colin, but since we are using it for multiple entries it was not ideal.  This is the simplest solution and I don't think it adds a lot of overhead, but happy to discuss/pair if there are concerns!

I didn't make a build of this in monolith, but confimed it fixes the homepage crashing if you turn off IntersectionObserver in Safari.  I did make a portal app build that fixes /catalog and any hub page (which all use the HorizontalScrollMenu)