fix(SocialMediaSharing): fix partially undefined id
Only prepends `sectionId-` to the svg title id if there is a `sectionId`. Previously, when a section id was not passed the title ids would look like `undefined-facebook`. Also adds in `displayName` prop in the `SOCIAL_SHARING_PLATFORMS` array of objects to be used for the dropdown variant in portal.