chore(deps): bump marked from 0.7.0 to 4.3.0
This pull request updates marked to version ^4.3.0 to fix current security vulnerabilities. The `parseInline()` function is used in place of `inlineLexer()`. 

Also deployed on portal and in the le in this [pr](https://github.com/codecademy-engineering/mono/pull/2819).

Please note in order for this to pass tests in monorepo, I needed to set the `mangle: false` in the markedOptions, in packages/gamut/src/Markdown/index.tsx. E

```
 const markedOptions = {
      smartypants: true,
      headerIds,
      headerPrefix: 'heading-',
      mangle: false,
    };
```
Mangle is mentioned in the documentation [here](https://marked.js.org/using_advanced) and is deprecated in the most recent version of marked in favor of the marked-mangle package. Mangling is used for mailto links to prevent email harvesting bots from collecting email addresses from websites. Currently, we are not mangling emails on prod.