fix(Input): use activatedStyle from props when value is provided
This PR adds back the ability to manually control an `Input`'s activated style by referring to the activated prop when it's passed in.

This change came about while I was working on an accessibility ticket around the recurly fields. While separating out the fields from 1 field to separate fields so that we could focus them individually on error, I noticed that the recurly inputs were starting in an activated state. For whatever reason Recurly is firing onChange events before the user interacts with the field. This makes `Input` think the field has been activated when it hasn't been interacted with yet. I noticed this prop existed but I think the functionality associated with the prop was removed. I'm not sure if that was intentional (it was a long time ago now) so I'm proposing to add it back in.