fix(FormGroupLabel): enable screenreaders to read a label's "optional" text
Currently, Gamut's forms' labels include either `*` to signify required, or `(optional)` to tell learners what fields are necessary. 
Both are set to `aria-hidden` but a test.io filing requests that the `(optional)` be read aloud.

This PR changes the `(optional)` text to be read along with the label text and keeps the `aria-hidden` on the `*`.