feat(Checkbox)!: require aria-label when label is ReactNode
Updates `Checkbox` to require an aria label if the label is a ReactNode. Updates `ConnectedCheckbox` to take an aria label and pass it down to `Checkbox`. Also updates `Checkbox` border radius to be hard-coded at 4px instead of using `md` token so that the border radius is not impacted for other themes that have different border radii tokens (i.e. LX studio).